requirements.txt
## Setup
# python3 -m venv .venv
# source .venv/bin/activate (on windows: .venv\Scripts\activate.bat)
# pip install -r requirements.txt

## Packages
# Google 
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Discordpy 
discord.py

# Other
asyncpg